filesystem: decode select response for MF
We always use the select response parser from the parent file. The usual
case will be that on the way up at least one parent has overloeaded the
response parser from the MF. However, this is not the case when an EF at
the top level is selected, then we will just get the hexdump of the
response since no response parser is available. Lets add a response
parser for the MF as well so that we always parse the select response.

Change-Id: Ib2b2476fc0e7d34cdacf7eb100ffadceb788c4d2
Related: OS#4963