filesystem: be more strict in method add_file()
The file identifier of a file is strictly defined as a two digit
hexadecimal number. Do not allow adding child files that violate this
constraint.

Since the human readable name also plays a major role in identifying
the type of the file (ADF, EF or DF) ensure that when adding files it
is ensured that the file actually has a name and that the name starts
with "DF.", "EF." or "ADF.".

Change-Id: I096907285b742e611d221b03ba067ea2522e7e52
Related: OS#4963